Update import paths and rename decorator functions
