+++ packages/server/src/common/decorator/index.ts
@@ -1,0 +1,1 @@
+export * from './api-docs.decorator'
