+++ packages/server/src/common/index.ts
@@ -1,2 +1,3 @@
+export * from './decorator'
+export * from './types'
 export * from './assert'
-export * from './type'
