+++ packages/server/src/common/types/index.ts
@@ -1,0 +1,2 @@
+export * from './api-docs.type'
+export * from './typia-validation.type'
